move tests to a separate file
